refactor: the returned data from the get all tables handler
